% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{generate_aggregate}
\alias{generate_aggregate}
\title{generate_aggregate}
\usage{
generate_aggregate(json_events_intermediate)
}
\arguments{
\item{json_events_intermediate}{a json string (see details below)}
}
\value{
a list with two elements: \code{"updated_intermediate"} and
\code{"aggregate"}.
}
\description{
Combines a previous intermediate object with the information contained
in new events to generate an updated intermediate object and the
corresponding aggregate list for plotting.
}
\details{
The json_events_intermediate is a json object that must contain
\code{"new_events"} (a json string) and \code{"intermediate"} (a json string).\cr
The \code{"new_events"} json string is an array of
objects (set of name/value pairs). The name/value pairs must contain the following
objects:
\itemize{
\item user,
\item name,
\item email,
\item date,
\item type,
\item percentage,
\item url,
\item title,
\item profile\cr
and has the following structure:\cr
\{"user":"22","name":"John Doe","title":"Unit 1: Phrasal verbs", ...\}\cr
The \code{"intermediate"} json string is a object (set of name/value pairs),
the names are "last_event", "visited_units", "users", "visitors",
"user_url_time" and "daily_effort".\cr
The function returns a list with two elements:\cr
\code{intermediate} is a list that contains the following objects\cr
\item \code{"last_event"}
\item \code{"visited_units"}
\item \code{"users"}
\item \code{"visitors"}
\item \code{"user_url_time"}
\item \code{"daily_effort"}
\code{aggregate} which is a list that contains the following objects
\item \code{"number_visited_units"}
\item \code{"number_users"}
\item \code{"users"}
\item \code{"visited_completed_units"}
\item \code{"user_url_visited_completed"}
\item \code{"user_url_time_percentage"}
\item \code{"percentage_user_wide"}
\item \code{"time_user_wide"}
\item \code{"objectives_quartiles"}
\item \code{"list_paths"}
\item \code{"daily_effort"}
}
}
