% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INDIeAnalyticsIncremental-package.R
\docType{package}
\name{INDIeAnalyticsIncremental-package}
\alias{INDIeAnalyticsIncremental}
\alias{INDIeAnalyticsIncremental-package}
\title{INDIeAnalyticsIncremental: The API to incrementally update the aggregate objects for INDIeAnalytics}
\description{
Contains the mail function generate_aggregate for the API to be served by a Docker container. It incrementally updates the aggregate objects to be used by the Shiny dashboard for INDIeAnalytics. To do so, it defines intermediate objects that are incrementally updated with new events' information, and transforms the intermediate objects into aggregate objects to be displayed in the dashboard.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
