% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utility_functions.R
\name{intermediate2aggregate}
\alias{intermediate2aggregate}
\title{intermediate2aggregate}
\usage{
intermediate2aggregate(intermediate)
}
\arguments{
\item{intermediate:}{a list with many objects, see details}
}
\value{
a list with the following objects:
\itemize{
\item \code{"number_visited_units"}
\item \code{"number_users"}
\item \code{"users"}
\item \code{"visited_completed_units"}
\item \code{"user_url_visited_completed"}
\item \code{"user_url_time_percentage"}
\item \code{"percentage_user_wide"}
\item \code{"time_user_wide"}
\item \code{"objectives_quartiles"}
\item \code{"list_paths"}
\item \code{"daily_effort"}
}
}
\description{
Computes, from an \code{"intermediate"} list, a list \code{"agregate"}
which contains the objects that can be directly used for plotting in shiny
without further manipulation.
}
\details{
The \code{"intermediate"} object is a list with the following objects:
\itemize{
\item \code{"last_event"}
\item \code{"visited_units"}
\item \code{"users"}
\item \code{"visitors"}
\item \code{"user_url_time"}
\item \code{"daily_effort"}
Those objects are intermediate objects which can be updated using the
information contained in new events, but must be transformed for plotting
Think for example that if a mean were to be plot, intermediate would
contain both the sum and the number of elements. These two quantities can
be updated upon arrival of new information, but the aggregated quantity
of interest is the mean.
}
}
