% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utility_functions.R
\name{update_intermediate}
\alias{update_intermediate}
\title{update_intermediate}
\usage{
update_intermediate(intermediate, new_events)
}
\arguments{
\item{intermediate:}{a list with many objects.}

\item{new_events:}{a dataframe, see details below.}
}
\value{
a list with the same structure than \code{intermediate} but where
the objects have been updated taking into account the information contained
in \code{"new_events"}
}
\description{
Updates the intermediate list that contains auxiliary objects.
The intermediate list  is used to produce the aggregate list.
It therefore needs to be updated combining the present intermediate object
with the new events that have been registered
}
\details{
The \code{"new_events"} dataframe should have been processed using the
\code{"prepare_events"} function. The latter adds in particular
\code{"time_spent"}  and \code{"duration_session"}, possibly taking into
account that the session spans from a previous call to the API
The \code{"intermediate"} is a list with the following objects:
\itemize{
\item \code{"last_event"}
\item \code{"visited_units"}
\item \code{"users"}
\item \code{"visitors"}
\item \code{"user_url_time"}
\item \code{"daily_effort"}
}
}
